requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.3
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
oauthlib==3.1.1
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.36
pickleshare==0.7.5
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
